% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/split_path.R
\name{split_path}
\alias{split_path}
\title{Split paths into vectors of folders and files}
\usage{
split_path(paths, drop = TRUE, ...)
}
\arguments{
\item{paths}{A vector of paths}

\item{drop}{If TRUE and \code{paths} is a single path,
then a vector and not a list of vectors is returned.}

\item{...}{Not used.}
}
\value{
Returns a list of character vectors, unless \code{drop}
is TRUE in case a single character vectors is returned.
}
\description{
Split paths into vectors of folders and files
}

